+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.5.4] - TBA
 
 - #154 added Paystone, Ashby rules (thanks @austinhuang0131)
+- #146 added Liveclicker & eGain, improved Zeta Global, Customer.io, Granicus rules (thanks @m-schmitt)
 - improved Zoho Campaigns rule, added YouTube
 - added Benchmark Email, Epsilon (thanks James M.)
 - more spacers identified as false positives (thanks James M.)